Split ScreenGrabTest into multiple tests. (#1470)
This has two advantages:
* If one of the test fails then the others can still run and we can get
  screenshots for them. Currently if ScreenGrabTest fails then all
  following screenshots are missing.
* Independent tests are more stable and easier to tweak.
